+++ com.github.fhd.eclipsecolortheme/src/com/github/fhd/eclipsecolortheme/ColorThemeManager.java
@@ -17,7 +17,8 @@ import com.github.fhd.eclipsecolortheme.mapper.WebEditorMapper.Type;
 public class ColorThemeManager {
     private static final String[] THEME_FILES = new String[] {
             "zenburn.xml",
-            "inkpot.xml"
+            "inkpot.xml",
+            "vibrantink.xml"
     };
     private Map<String, Map<String, String>> themes;
     private Set<ThemePreferenceMapper> editors;