+++ CHANGES.md
@@ -3,6 +3,7 @@
 * Added support for the C++ editor.
 * Moved colour themes to XML files.
 * Fixed mark occurrence highlighting.
+* Added the Vibrant Ink theme.
 
 0.3.1 (2011-01-09)
 ==================